Updated everything that's related to screen position to use floats
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@1107 3db68df2-c116-0410-a063-a993310a9797

+++ source/gui/CCheckBox.cpp
@@ -113,7 +113,7 @@ void CCheckBox::Draw()
 
 	// Get square
 	// TODO Gee: edit below when CRect has got "height()"
-	int middle = (m_CachedActualSize.bottom - m_CachedActualSize.top)/2;
+	float middle = (m_CachedActualSize.bottom - m_CachedActualSize.top)/2;
 	CRect rect;
 	rect.left =		m_CachedActualSize.left + middle - square_side/2;
 	rect.right =	rect.left + square_side;
